+++ packages/components/pie-checkbox-group/src/defs.ts
@@ -14,6 +14,11 @@ export interface CheckboxGroupProps {
     label?: string;
 
     /**
+     * Inline (horizontal) positioning of checkbox items
+     */
+    isInline?: boolean;
+
+    /**
      * An optional assistive text to display below the checkbox group.
      */
     assistiveText?: string;
@@ -35,10 +40,12 @@ export interface CheckboxGroupProps {
  * @constant
  */
 export const ON_CHECKBOX_GROUP_DISABLED = 'pie-checkbox-group-disabled';
+export const ON_CHECKBOX_GROUP_ERROR = 'pie-checkbox-group-error';
 
-export type DefaultProps = ComponentDefaultProps<CheckboxGroupProps, 'status' | 'disabled'>;
+export type DefaultProps = ComponentDefaultProps<CheckboxGroupProps, 'status' | 'disabled' | 'isInline'>;
 
 export const defaultProps: DefaultProps = {
     status: 'default',
     disabled: false,
+    isInline: false,
 };