+++ packages/components/pie-toast/src/defs.ts
@@ -1,5 +1,4 @@
 import { type ComponentDefaultProps } from '@justeattakeaway/pie-webc-core';
-
 export const variants = ['neutral', 'info', 'warning', 'success', 'error'] as const;
 
 // TODO - please remove the eslint disable comment below when you add props to this interface
@@ -21,7 +20,6 @@ export interface ToastProps {
    * When true, the toast is set to be open and visible.
    */
   isOpen?: boolean;
-
   /**
    * Sets the variant of the toast.
    */
@@ -31,7 +29,6 @@ export interface ToastProps {
    * When true, the toast is displayed with a strong visual style.
    */
   isStrong?: boolean;
-
   /**
    * When true, allows dismissing the toast by clicking on the close button.
    */
@@ -77,8 +74,8 @@ export const ON_TOAST_LEADING_ACTION_CLICK_EVENT = `${componentSelector}-leading
 export type DefaultProps = ComponentDefaultProps<ToastProps, 'isOpen' | 'isDismissible' | 'isMultiline' | 'variant' | 'isStrong'>;
 
 export const defaultProps: DefaultProps = {
-    variant: 'neutral',
     isOpen: true,
+    variant: 'neutral',
     isStrong: false,
     isDismissible: true,
     isMultiline: false,