+++ apps/examples/wc-angular12/CHANGELOG.md
@@ -1,5 +1,23 @@
 # wc-angular12
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [[`927b800ed`](https://github.com/justeattakeaway/pie/commit/927b800ed2456c6db299694fbf850dea4aed34bd)]:
+  - @justeattakeaway/pie-webc@0.5.15
+
+## 0.5.0
+
+### Minor Changes
+
+- [Added] pie-webc integration to angular examples app ([#1611](https://github.com/justeattakeaway/pie/pull/1611)) by [@kevinrodrigues](https://github.com/kevinrodrigues)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.5.14
+
 ## 0.4.43
 
 ### Patch Changes